test: Refactor RSpec OAuth helper functions to own file
