+++ spec/requests/temporary_oauth_sessions_spec.rb
@@ -1,4 +1,9 @@
 require 'rails_helper'
+require './spec/support/oauth_helpers'
+
+RSpec.configure do |c|
+  c.include Helpers::OAuth
+end
 
 RSpec.describe "Temporary OAuth requests", type: :request do
   before do
@@ -102,35 +107,6 @@ RSpec.describe "Temporary OAuth requests", type: :request do
       end
     end
   end
-end
-
-def oauth_username_to_mock_uid(provider, username)
-  "#{provider}|#{username}"
-end
-
-def stub_oauth_flow(provider, username, &block)
-  base_mock = {
-    uid: oauth_username_to_mock_uid(provider, username),
-    provider: provider,
-    info: {
-      provider == "bnet" ? :battletag : :name => username,
-      image: "https://ehe.gg/media/img/logos/Elo-Hell-Logo_I-C-Dark.png"
-    }
-  }
-  # Discord-specific adjustments
-  # ! If needed in the future for other services,
-  # ! can refactor adjustments into a dedicated method
-  if provider == "discord"
-    base_mock[:info][:name] = username.split("#").first
-    base_mock[:extra] = {
-      raw_info: {
-        discriminator: username.split("#").last
-      }
-    }
-  end
-  OmniAuth.config.mock_auth[provider.to_sym] = OmniAuth::AuthHash.new(base_mock)
-  block.call
-  OmniAuth.config.mock_auth[provider.to_sym] = nil
 end
 
 def obtain_temporary_oauth_session
